Refactor test main out of unit/
