+++ test/main.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+
+#include "minunit.h"
+#include "unit/test.c"
+
+int main(int argc, char **argv) {
+    run_all_unit_tests();
+}
